Isolate thresholds list
